chore: Remove unused fields from Climb object
This object is a smaller subset of all the fields in a full climb object. It is only intended for displaying in the area screen list of climbs. A new full climb object will be made that will include all details of climbs for when you want to view a particular climb, but all this information is unnecessary when fetching an area and would load to a very large response body.